mid refactor
So, er, I realise I'm doing a lot of `public static` like I want to be
doing FP rather than OO. That's fine for sketching out algorithms, but
I'm getting up to speed with an OO language, so let's start to actually
engage with the prevailing idioms...

Refactoring into OCR (the line) and OCRDigit classes.
Having done TDD throughout, I was unsure exactly how I wanted to
structure this, and have just hacked at the source code instead of
test-first. On reflection, this would be the ideal place to do test
first, to sketch out the interface.

Oh well, good learning experience I guess. Let's call it a "spike" this
time, and remember to do better next time.

Have taken it to a point where the obvious compiler errors are fixed in
src, and the errors are now reported from the Tests, so next I'll
rewrite those till I'm sure the structure works, and then fix remaining
issues with code.